put mongo into dataBaseConnection and created there IsDbConnected; put 3 methods of DataBaseService into index; created by import priorities in handlers; and removed clientConnectionController.
