+++ src/services/DataBaseService/dataBaseConnection.ts
@@ -1,0 +1,27 @@
+import mongoose from 'mongoose';
+
+export class DataBaseConnection {
+  private isDbConnected: boolean;
+
+  constructor() {
+    this.isDbConnected = false;
+  }
+
+  public async connectMongodb() {
+    try {
+      const dburl = process.env.MONGODB_URL;
+
+      if (!dburl) {
+        throw new Error('MONGODB_URL environment variable is not defined');
+      }
+
+      await mongoose.connect(dburl);
+      this.isDbConnected = true;
+    } catch (err) {
+      console.error('Error connecting to MongoDB:', err);
+    }
+  }
+  public getisDbConnected(): boolean {
+    return this.isDbConnected;
+  }
+}
